pyproject.toml
[tool.poetry]
name = "academic_paper_database"
version = "0.1.0"
description = "A database to store, search, and categorize academic research papers."
authors = ["Marshall Bruner <brunerm99@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.14"
click = "^7.1.2"
dash = "^1.16.1"
markupsafe = "^2.0.1"
Flask = "^1.1.2"
Flask-Caching = "^1.9.0"
Flask-Login = "^0.5.0"
Flask-Migrate = "^2.5.3"
Flask-SQLAlchemy = "^2.4.4"
Flask-Testing = "^0.8.0"
Flask-WTF = "^0.14.3"
gevent = "^20.6.2"
greenlet = "^0.4.16"
gunicorn = "^20.0.4"
numpy = "^1.19.2"
pandas = "^1.1.2"
psycopg2-binary = "^2.8.6"
python-dateutil = "^2.8.1"
python-dotenv = "^0.14.0"
python-editor = "^1.0.4"
pytz = "^2020.1"
PyYAML = "^5.3.1"
dash-bootstrap-components = "^0.10.6"
scikit-learn = "^0.23.2"
ipykernel = "^5.4.2"
matplotlib = "^3.3.3"
keras = "^2.4.3"
statsmodels = "^0.12.1"
jwt = "^1.1.0"
flask-bootstrap4 = "^4.0.2"
MarkupSafe = "2.0.1"
tabulate = "^0.8.10"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"